+++ libs/angular/src/lib/store-signals.type.ts
@@ -1,0 +1,8 @@
+import { Selectors } from '@state-adapt/core';
+
+export type StoreSignals<State, S extends Selectors<State>> = {
+  (): State;
+  readOnce: () => State;
+} & {
+  [K in keyof S]: () => ReturnType<S[K]>;
+};