+++ libs/rxjs/src/index.ts
@@ -17,6 +17,7 @@ export * from './lib/http/get-http-sources.function';
 
 export * from './lib/stores/selections.type';
 export * from './lib/stores/sources.type';
+export * from './lib/stores/core-store-props.type';
 export * from './lib/stores/smart-store.interface';
 export * from './lib/stores/join-stores.function';
 export * from './lib/stores/store-like.type';