+++ libs/rxjs/src/lib/stores/joined-store.interface.ts
@@ -1,13 +1,6 @@
-import { Observable } from 'rxjs';
 import { AnySelectors } from '@state-adapt/core';
 import { Selections } from '../stores/selections.type';
+import { WithCoreStoreProps } from './core-store-props.type';
 
-export type JoinedStore<State, S1 extends AnySelectors> = Selections<State, S1> & {
-  __: {
-    requireSources$: Observable<any>;
-    fullSelectors: S1;
-    selectors: S1;
-    initialState: State;
-    select: <State>(sel: any) => Observable<State>;
-  };
-};
+export type JoinedStore<State, S extends AnySelectors> = Selections<State, S> &
+  WithCoreStoreProps<State, S>;