+++ libs/rxjs/src/lib/stores/smart-store.interface.ts
@@ -1,17 +1,12 @@
 import { Selectors } from '@state-adapt/core';
-import { Observable } from 'rxjs';
 import { Selections } from '../stores/selections.type';
+import { CoreStoreProps } from './core-store-props.type';
 
 export type SmartStore<State, S extends Selectors<State>> = Selections<State, S> & {
   /**
    * Don't use this property directly. Intended for internal use only.
    */
-  __: {
-    requireSources$: Observable<any>;
-    fullSelectors: S;
-    selectors: S;
-    initialState: State;
+  __: CoreStoreProps<State, S> & {
     path: string;
-    select: <State>(sel: any) => Observable<State>;
   };
 };